adicionando metodo para retornar lista de tickets fechados por mes

+++ src/main/java/br/ufma/glp/unidesk/backend/api/v1/dto/model/DashboardModel.java
@@ -17,7 +17,7 @@ public class DashboardModel {
     Long totalTicketsPendentes;
     Long totalTicketsEmAndamento;
     Long porcentagemProgresso;
-    Long porcentagemAbertos;
+    Double porcentagemAbertos;
     Long porcentagemResolvidos;
     Long porcentagemAndamento;
 }
